Moved the drop down and made it display if the user already has an auth
The drop down menu was moved from next to the username to the bottom of
the comment.

The drop down items are now coloured. They are grey if the user has no
auth for that script, green if they have an auth and orange if they have
an auth but it has expired.